Remaking underlying Vec library.
I think it should be a class, because that allows `a.plus(b)` which is a lot more readable than `plusVec(a, b)`.

Also, it means that the SolidStore system won't start proxying the internals, which I don't think it should be doing.

I don't export it yet, but I'll do it when I work up the courage to change Marble Marcher